修改了亮度值为浮点数

+++ i2c.h
@@ -17,7 +17,7 @@ void Delay_ms(unsigned long nValue);
 void Delay_us(unsigned long nValue);
 
 void InitALS(void);
-int GetALSValue();
+float GetALSValue();
 
 void test(void);
